Add Gemini exchange

+++ src/jvmMain/kotlin/com/run3wide/bitness/pricing/service/ExchangeRateService.kt
@@ -11,6 +11,6 @@ class ExchangeRateService(
 
     fun getBtcExchangeRate(): ExchangeRateDto {
         return exchangeRateRepository.getBtcExchangeRate()
-            .let { ExchangeRateDto(it.amount, it.currency) }
+            .let { ExchangeRateDto(it.amount.toDouble(), it.currency) }
     }
 }
